--- .github/workflows/secrets.test-memory-ai5.yml
@@ -1,166 +1,0 @@
-name: Memory Tests (StreamVNext/AI5)
-
-on:
-  workflow_run:
-    workflows: ['Quality assurance']
-    types:
-      - completed
-
-jobs:
-  check-changes:
-    if: ${{ github.repository == 'mastra-ai/mastra' }}
-    runs-on: ubuntu-latest
-    outputs:
-      memory-changed: ${{ steps.changes.outputs.memory }}
-    permissions:
-      contents: read
-      statuses: write
-    steps:
-      - uses: actions/checkout@v5
-        with:
-          fetch-depth: 0
-          ref: ${{ github.event.workflow_run.head_sha }}
-
-      - name: Set pending status
-        uses: ./.github/workflows/shared-actions/set-pr-status
-        with:
-          status: 'pending'
-          context: 'Memory Tests (StreamVNext/AI5)'
-          description: 'Checking for changes'
-          sha: ${{ github.event.workflow_run.head_sha }}
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
-
-      - name: Check for memory package changes
-        uses: dorny/paths-filter@v3
-        id: changes
-        with:
-          base: main
-          ref: ${{ github.event.workflow_run.head_sha }}
-          filters: |
-            memory:
-              - '!(**/*.md)packages/**'
-              - '!(**/*.md)stores/**'
-              - '!(**/*.md)client-sdks/**'
-
-  skip-tests:
-    needs: check-changes
-    if: needs.check-changes.outputs.memory-changed == 'false'
-    runs-on: ubuntu-latest
-    permissions:
-      statuses: write
-    steps:
-      - uses: actions/checkout@v5
-      - name: Set success status for unchanged memory
-        uses: ./.github/workflows/shared-actions/set-pr-status
-        with:
-          status: 'success'
-          context: 'Memory Tests (StreamVNext/AI5)'
-          description: 'Memory packages unchanged - skipping tests'
-          sha: ${{ github.event.workflow_run.head_sha }}
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
-
-  test:
-    needs: check-changes
-    if: ${{ github.repository == 'mastra-ai/mastra' && needs.check-changes.outputs.memory-changed == 'true' }}
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-    timeout-minutes: 10
-    strategy:
-      fail-fast: false
-      matrix:
-        suite:
-          - streaming
-          - working-memory
-          - agent-memory
-          - processors
-          - pg
-    services:
-      postgres:
-        image: pgvector/pgvector:pg16
-        env:
-          POSTGRES_USER: postgres
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: mastra
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-
-      redis:
-        image: redis
-        ports:
-          - 6379:6379
-
-    env:
-      TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
-      TURBO_TEAM: ${{ secrets.TURBO_TEAM }}
-      TURBO_CACHE: remote:r
-
-    steps:
-      - uses: actions/checkout@v5
-        with:
-          ref: ${{ github.event.workflow_run.head_sha }}
-
-      - name: Setup pnpm and Node.js
-        uses: ./.github/actions/setup-pnpm-node
-
-      - name: Build vector+store packages
-        run: pnpm turbo build --filter "./stores/*" --filter "mastra" --filter "@mastra/memory" --filter "@mastra/fastembed"
-
-      - name: Install integration test dependencies
-        run: pnpm install --ignore-workspace
-        working-directory: packages/memory/integration-tests-v5
-
-      - name: Run Memory ${{ matrix.suite }} tests
-        run: pnpm run test:${{ matrix.suite }}
-        working-directory: packages/memory/integration-tests-v5
-        env:
-          NODE_OPTIONS: '--max_old_space_size=8096'
-          DB_URL: 'postgresql://postgres:postgres@localhost:5432/mastra'
-          OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
-          OPENROUTER_API_KEY: ${{ secrets.OPENROUTER_API_KEY }}
-          KV_REST_API_URL: ${{ secrets.KV_REST_API_URL }}
-          KV_REST_API_TOKEN: ${{ secrets.KV_REST_API_TOKEN }}
-          GOOGLE_GENERATIVE_AI_API_KEY: ${{ secrets.GOOGLE_GENERATIVE_AI_API_KEY }}
-
-  test-success:
-    needs: [check-changes, test]
-    if: ${{ always() && needs.check-changes.outputs.memory-changed == 'true' && needs.test.result == 'success' }}
-    runs-on: ubuntu-latest
-    permissions:
-      statuses: write
-    steps:
-      - uses: actions/checkout@v5
-      - name: Set success status for completed tests
-        uses: ./.github/workflows/shared-actions/set-pr-status
-        with:
-          status: 'success'
-          context: 'Memory Tests (StreamVNext/AI5)'
-          description: 'All memory tests passed'
-          sha: ${{ github.event.workflow_run.head_sha }}
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
-
-  test-failure:
-    needs: [check-changes, test]
-    if: ${{ always() && needs.check-changes.outputs.memory-changed == 'true' && needs.test.result == 'failure' }}
-    runs-on: ubuntu-latest
-    permissions:
-      statuses: write
-    steps:
-      - uses: actions/checkout@v5
-      - name: Set failure status for failed tests
-        uses: ./.github/workflows/shared-actions/set-pr-status
-        with:
-          status: 'failure'
-          context: 'Memory Tests (StreamVNext/AI5)'
-          description: 'One or more memory tests failed'
-          sha: ${{ github.event.workflow_run.head_sha }}
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
