+++ .github/workflows/secrets.test-memory.yml
@@ -2,7 +2,7 @@ name: Memory Tests
 
 on:
   workflow_run:
-    workflows: ['Quality assurance']
+    workflows: ["Quality assurance"]
     types:
       - completed
 
@@ -24,9 +24,9 @@ jobs:
       - name: Set pending status
         uses: ./.github/workflows/shared-actions/set-pr-status
         with:
-          status: 'pending'
-          context: 'Memory Tests'
-          description: 'Checking for changes'
+          status: "pending"
+          context: "Memory Tests"
+          description: "Checking for changes"
           sha: ${{ github.event.workflow_run.head_sha }}
           github_token: ${{ secrets.GITHUB_TOKEN }}
           target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
@@ -54,9 +54,9 @@ jobs:
       - name: Set success status for unchanged memory
         uses: ./.github/workflows/shared-actions/set-pr-status
         with:
-          status: 'success'
-          context: 'Memory Tests'
-          description: 'Memory packages unchanged - skipping tests'
+          status: "success"
+          context: "Memory Tests"
+          description: "Memory packages unchanged - skipping tests"
           sha: ${{ github.event.workflow_run.head_sha }}
           github_token: ${{ secrets.GITHUB_TOKEN }}
           target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
@@ -133,12 +133,13 @@ jobs:
         run: pnpm run test:${{ matrix.suite }}
         working-directory: packages/memory/integration-tests
         env:
-          NODE_OPTIONS: '--max_old_space_size=8096'
-          DB_URL: 'postgresql://postgres:postgres@localhost:5432/mastra'
+          NODE_OPTIONS: "--max_old_space_size=8096"
+          DB_URL: "postgresql://postgres:postgres@localhost:5432/mastra"
           OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
           KV_REST_API_URL: ${{ secrets.KV_REST_API_URL }}
           KV_REST_API_TOKEN: ${{ secrets.KV_REST_API_TOKEN }}
           GOOGLE_GENERATIVE_AI_API_KEY: ${{ secrets.GOOGLE_GENERATIVE_AI_API_KEY }}
+          OPENROUTER_API_KEY: ${{ secrets.OPENROUTER_API_KEY }}
 
   test-success:
     needs: [check-changes, test]
@@ -151,9 +152,9 @@ jobs:
       - name: Set success status for completed tests
         uses: ./.github/workflows/shared-actions/set-pr-status
         with:
-          status: 'success'
-          context: 'Memory Tests'
-          description: 'All memory tests passed'
+          status: "success"
+          context: "Memory Tests"
+          description: "All memory tests passed"
           sha: ${{ github.event.workflow_run.head_sha }}
           github_token: ${{ secrets.GITHUB_TOKEN }}
           target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
@@ -169,9 +170,9 @@ jobs:
       - name: Set failure status for failed tests
         uses: ./.github/workflows/shared-actions/set-pr-status
         with:
-          status: 'failure'
-          context: 'Memory Tests'
-          description: 'One or more memory tests failed'
+          status: "failure"
+          context: "Memory Tests"
+          description: "One or more memory tests failed"
           sha: ${{ github.event.workflow_run.head_sha }}
           github_token: ${{ secrets.GITHUB_TOKEN }}
           target_url: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
