--- packages/memory/integration-tests-v5/.env.test
@@ -1,5 +1,0 @@
-# PostgreSQL connection
-DB_URL=postgres://postgres:password@localhost:5434/mastra
-
-# OpenAI (for vector embeddings)
-# OPENAI_API_KEY=your_openai_api_key
