--- packages/memory/integration-tests-v5/.gitignore
@@ -1,4 +1,0 @@
-memory.db
-*.db
-*.db*
-src/worker/*.js
