--- packages/memory/integration-tests-v5/docker-compose.yml
@@ -1,39 +1,0 @@
-services:
-  postgres:
-    image: ankane/pgvector:v0.5.1
-    ports:
-      - '5434:5432'
-    environment:
-      POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: password
-      POSTGRES_DB: mastra
-    volumes:
-      - pg_data:/var/lib/postgresql/data
-    healthcheck:
-      test: ['CMD-SHELL', 'pg_isready -U postgres']
-      interval: 5s
-      timeout: 5s
-      retries: 5
-  redis:
-    image: redis:8-alpine
-    ports:
-      - '6370:6379'
-    command: redis-server --requirepass redis_password
-    healthcheck:
-      test: ['CMD', 'redis-cli', '--no-auth-warning', '-a', 'redis_password', 'ping']
-      interval: 5s
-      timeout: 5s
-      retries: 5
-  serverless-redis-http:
-    image: hiett/serverless-redis-http:latest
-    ports:
-      - '8079:80'
-    environment:
-      SRH_MODE: env
-      SRH_TOKEN: test_token
-      SRH_CONNECTION_STRING: 'redis://:redis_password@redis:6379'
-    depends_on:
-      redis:
-        condition: service_healthy
-volumes:
-  pg_data:
