--- packages/memory/integration-tests-v5/package.json
@@ -1,62 +1,0 @@
-{
-  "name": "@mastra/memory-integration-tests-v5",
-  "private": true,
-  "version": "0.1.10-alpha.1",
-  "scripts": {
-    "test": "pnpm test:streaming && pnpm test:working-memory && pnpm test:working-memory-additive && pnpm test:agent-memory && pnpm test:processors && pnpm test:fastembed",
-    "test:agent-memory": "vitest run ./src/agent-memory.test.ts",
-    "test:libsql": "vitest run ./src/with-libsql-storage.test.ts",
-    "test:pg": "vitest run ./src/with-pg-storage.test.ts",
-    "test:streaming": "vitest run ./src/streaming-memory.test.ts",
-    "test:working-memory": "vitest run ./src/working-memory.test.ts",
-    "test:working-memory-additive": "vitest run ./src/working-memory-additive.test.ts",
-    "test:processors": "vitest run ./src/processors.test.ts",
-    "test:fastembed": "vitest run ./src/fastembed.test.ts",
-    "test:ordering": "vitest run ./src/message-ordering.test.ts",
-    "test:ai-sdk-ids": "vitest run ./src/ai-sdk-duplicate-ids.test.ts",
-    "dev": "mastra dev"
-  },
-  "dependencies": {
-    "@mastra/ai-sdk": "latest",
-    "@mastra/client-js": "latest",
-    "@mastra/fastembed": "latest",
-    "@mastra/libsql": "latest",
-    "@mastra/memory": "latest",
-    "@mastra/pg": "latest",
-    "dotenv": "^16.4.7",
-    "json-schema": "^0.4.0",
-    "zod": "^3.25.76"
-  },
-  "devDependencies": {
-    "@ai-sdk/anthropic": "latest",
-    "@ai-sdk/google": "latest",
-    "@ai-sdk/openai": "^2.0.15",
-    "@ai-sdk/provider": "^2.0.0",
-    "@ai-sdk/react": "^2.0.15",
-    "@faker-js/faker": "^9.9.0",
-    "@mastra/core": "latest",
-    "@testing-library/react": "^16.3.0",
-    "@types/json-schema": "^7.0.15",
-    "@types/node": "22.13.17",
-    "ai": "latest",
-    "jsdom": "^26.1.0",
-    "mastra": "latest",
-    "next": "^15.5.7",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "typescript": "^5.8.2",
-    "vitest": "^4.0.8"
-  },
-  "pnpm": {
-    "overrides": {
-      "@mastra/ai-sdk": "link:../../../client-sdks/ai-sdk",
-      "@mastra/client-js": "link:../../../client-sdks/client-js",
-      "@mastra/core": "link:../../core",
-      "mastra": "link:../../cli",
-      "@mastra/memory": "link:../../memory",
-      "@mastra/fastembed": "link:../../fastembed",
-      "@mastra/libsql": "link:../../../stores/libsql",
-      "@mastra/pg": "link:../../../stores/pg"
-    }
-  }
-}
