+++ packages/memory/integration-tests/src/v6/mastra/tools/weather.ts
